(refactor) Refactor modals to match naming conventions
