+++ packages/esm-patient-registration-app/src/widgets/delete-identifier-confirmation.scss
