+++ packages/esm-patient-registration-app/translations/en.json
@@ -14,12 +14,15 @@
   "codedPersonAttributeNoAnswerSet": "The person attribute field '{{codedPersonAttributeFieldId}}' is of type 'coded' but has been defined without an answer concept set UUID. The 'answerConceptSetUuid' key is required.",
   "configure": "Configure",
   "configureIdentifiers": "Configure identifiers",
+  "confirmIdentifierDeletionText": "Are you sure you want to remove this identifier?",
   "contactSection": "Contact Details",
   "createNewPatient": "Create new patient",
   "dateOfBirthLabelText": "Date of Birth",
   "deathDateInputLabel": "Date of Death",
   "deathdayNotInTheFuture": "Death day cannot be in the future",
   "deathSection": "Death Info",
+  "deleteIdentifierModalHeading": "Remove identifier?",
+  "deleteIdentifierModalText": " has a value of ",
   "deleteIdentifierTooltip": "Delete",
   "deleteRelationshipTooltipText": "Delete",
   "demographicsSection": "Basic Info",
@@ -74,6 +77,7 @@
   "relationshipToPatient": "Relationship to patient",
   "relativeFullNameLabelText": "Related person",
   "relativeNamePlaceholder": "Firstname Familyname",
+  "removeIdentifierButton": "Remove Identifier",
   "resetIdentifierTooltip": "Reset",
   "restoreRelationshipActionButton": "Undo",
   "searchAddress": "Search address",
