+++ packages/esm-service-queues-app/src/add-patient-toqueue/add-patient-toqueue-dialog.component.tsx
@@ -11,17 +11,17 @@ import {
   SelectItem,
   RadioButtonGroup,
   RadioButton,
+  RadioButtonSkeleton,
+  SelectSkeleton,
 } from '@carbon/react';
 import { showSnackbar, useConfig } from '@openmrs/esm-framework';
 import { postQueueEntry } from '../active-visits/active-visits-table.resource';
-import styles from './add-patient-toqueue-dialog.scss';
-import { type ActiveVisit, useMissingQueueEntries } from '../visits-missing-inqueue/visits-missing-inqueue.resource';
+import { type ActiveVisit } from '../visits-missing-inqueue/visits-missing-inqueue.resource';
 import { useQueueLocations } from '../patient-search/hooks/useQueueLocations';
 import { useQueues } from '../hooks/useQueues';
 import { useMutateQueueEntries } from '../hooks/useQueueEntries';
 import { type ConfigObject } from '../config-schema';
-import { RadioButtonSkeleton } from '@carbon/react';
-import { SelectSkeleton } from '@carbon/react';
+import styles from './add-patient-toqueue-dialog.scss';
 
 interface AddVisitToQueueDialogProps {
   visitDetails: ActiveVisit;
