+++ packages/esm-service-queues-app/src/queue-table/queue-entry-actions/end-queue-entry.modal.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { useTranslation } from 'react-i18next';
 import { type QueueEntry } from '../../types';
 import { updateQueueEntry } from './queue-entry-actions.resource';
-import QueueEntryConfirmActionModal from './queue-entry-confirm-action-modal.component';
+import QueueEntryConfirmActionModal from './queue-entry-confirm-action.modal';
 
 interface EndQueueEntryModalProps {
   queueEntry: QueueEntry;
