+++ packages/esm-service-queues-app/src/queue-table/queue-entry-actions/queue-entry-confirm-action.test.tsx
@@ -4,9 +4,9 @@ import { mockQueues, mockQueueEntryAlice } from '__mocks__';
 import React from 'react';
 import { renderWithSwr } from 'tools';
 import userEvent from '@testing-library/user-event';
-import UndoTransitionQueueEntryModal from './undo-transition-queue-entry-modal.component';
-import VoidQueueEntryModal from './void-queue-entry-modal.component';
-import EndQueueEntryModal from './end-queue-entry-modal.component';
+import UndoTransitionQueueEntryModal from './undo-transition-queue-entry.modal';
+import VoidQueueEntryModal from './void-queue-entry.modal';
+import EndQueueEntryModal from './end-queue-entry.modal';
 
 const mockedOpenmrsFetch = openmrsFetch as jest.Mock;
