+++ packages/esm-service-queues-app/src/queue-table/queue-entry-actions/transition-queue-entry.modal.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { useTranslation } from 'react-i18next';
 import { type QueueEntry } from '../../types';
-import QueueEntryActionModal from './queue-entry-actions-modal.component';
+import QueueEntryActionModal from './queue-entry-actions.modal';
 import { transitionQueueEntry } from './queue-entry-actions.resource';
 import { convertTime12to24 } from '../../helpers/time-helpers';
