+++ packages/esm-service-queues-app/src/queue-table/queue-entry-actions/undo-transition-queue-entry.modal.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { useTranslation } from 'react-i18next';
 import { type QueueEntry } from '../../types';
 import { undoTransition } from './queue-entry-actions.resource';
-import QueueEntryConfirmActionModal from './queue-entry-confirm-action-modal.component';
+import QueueEntryConfirmActionModal from './queue-entry-confirm-action.modal';
 
 interface UndoTransitionQueueEntryModalProps {
   queueEntry: QueueEntry;
